Move OJS submission JSON parsing out of the model
